+++ app/src/androidTest/java/com/epicodus/myrestaurants/MainActivityInstrumentationTest.java
@@ -2,6 +2,8 @@ package com.epicodus.myrestaurants;
 
 import android.support.test.rule.ActivityTestRule;
 
+import com.epicodus.myrestaurants.ui.MainActivity;
+
 import org.junit.Rule;
 import org.junit.Test;
