+++ app/src/androidTest/java/com/epicodus/myrestaurants/RestaurantsActivityInstrumentationTest.java
@@ -3,6 +3,8 @@ package com.epicodus.myrestaurants;
 import android.support.test.rule.ActivityTestRule;
 import android.view.View;
 
+import com.epicodus.myrestaurants.ui.RestaurantsActivity;
+
 import org.junit.Rule;
 import org.junit.Test;
