+++ app/src/main/java/com/epicodus/myrestaurants/models/Restaurant.java
@@ -1,4 +1,4 @@
-package com.epicodus.myrestaurants;
+package com.epicodus.myrestaurants.models;
 import java.util.ArrayList;
 
 public class Restaurant {
