+++ app/src/main/java/com/epicodus/myrestaurants/ui/MainActivity.java
@@ -1,4 +1,4 @@
-package com.epicodus.myrestaurants;
+package com.epicodus.myrestaurants.ui;
 
 import android.content.Intent;
 import android.graphics.Typeface;
@@ -8,6 +8,8 @@ import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.TextView;
+
+import com.epicodus.myrestaurants.R;
 
 import butterknife.Bind;
 import butterknife.ButterKnife;
