+++ app/src/main/res/layout/activity_main.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.epicodus.myrestaurants.MainActivity"
+    tools:context="com.epicodus.myrestaurants.ui.MainActivity"
     android:background="#000000">
 
     <ImageView
