+++ app/src/main/res/layout/activity_restaurants.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.epicodus.myrestaurants.RestaurantsActivity">
+    tools:context="com.epicodus.myrestaurants.ui.RestaurantsActivity">
 
     <TextView
         android:layout_width="wrap_content"
