+++ app/src/test/java/com/epicodus/myrestaurants/MainActivityTest.java
@@ -4,6 +4,9 @@ import android.content.Intent;
 import android.os.Build;
 import android.widget.TextView;
 
+import com.epicodus.myrestaurants.ui.MainActivity;
+import com.epicodus.myrestaurants.ui.RestaurantsActivity;
+
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
